implemented bet spread and min bet increment

+++ src/main/java/com/arthur/blackjack/analytics/Analytics.java
@@ -3,7 +3,7 @@ package com.arthur.blackjack.analytics;
 public interface Analytics {
     void createNewResultsSheet(int bet);
     void recordNewRound(Integer round, Integer money);
-    void recordInitialBet(Integer round, Integer bet);
+    void recordInitialBet(Integer round, float bet);
     void recordRoundResult(Integer round, RoundResult result);
     void evaluateFormulas();
     void saveExcel();
