+++ src/main/java/com/arthur/blackjack/analytics/impl/CsvAnalyticsImpl.java
@@ -51,7 +51,7 @@ public class CsvAnalyticsImpl implements Analytics {
     }
 
     @Override
-    public void recordInitialBet(Integer round, Integer bet) {
+    public void recordInitialBet(Integer round, float bet) {
         writer.print(bet + ",");
     }
