+++ src/main/java/com/arthur/blackjack/models/hand/PlayerHand.java
@@ -1,6 +1,6 @@
 package com.arthur.blackjack.models.hand;
 
 public interface PlayerHand extends Hand {
-    Integer getBet();
-    void setBet(Integer bet);
+    float getBet();
+    void setBet(float bet);
 }
